+++ pkg/sql/zone_config.go
@@ -92,7 +92,7 @@ func getZoneConfig(
 }
 
 // ZoneConfigHook returns the zone config for the object with id using the
-// cached system config. If keySuffix is within a subzone, the subzone's config
+// cached system config. If keySuffix is within a subzone, the subzone's container
 // is returned instead.
 func ZoneConfigHook(
 	cfg config.SystemConfig, id uint32, keySuffix []byte,
