+++ pkg/storage/closedts/storage/storage_mem.go
@@ -12,7 +12,7 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package ctstorage
+package storage
 
 import (
 	"bytes"
